==> Lessons/09_DataVisualization.Rmd <==
---
title: "9: Data Visualization"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2019"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## LESSON OBJECTIVES
1. Perform simple data visualizations in the R package `ggplot`
2. Develop skills to adjust aesthetics and layers in graphs
3. Apply a decision tree framework for appropriate graphing methods

## SET UP YOUR DATA ANALYSIS SESSION

```{r, message = FALSE, warning = FALSE}
getwd()
library(tidyverse)

PeterPaul.chem.nutrients <- read.csv("./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv")
PeterPaul.nutrients.gathered <- read.csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed.csv")
PeterPaul.chem.nutrients.summaries <- read_csv("./Data/Processed/NTL-LTER_Lake_Summaries_PeterPaul_Processed.csv")
EPAair <- read.csv("./Data/Processed/EPAair_O3PM25_3sites1718_processed.csv")
dim(EPAair)

EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")
PeterPaul.chem.nutrients$sampledate <- as.Date(PeterPaul.chem.nutrients$sampledate, format = "%Y-%m-%d")

```

## GGPLOT

ggplot, called from the package `ggplot2`, is a graphing and image generation tool in R. This package is part of tidyverse. While base R has graphing capabilities, ggplot has the capacity for a wider range and more sophisticated options for graphing. 
ggplot has only a few rules: 

* The first line of ggplot code always starts with `ggplot()`
* A data frame must be specified within the `ggplot()` function. Additional datasets can be specified in subsequent layers.
* Aesthetics must be specified, most commonly x and y variables but including others. Aesthetics can be specified in the `ggplot()` function or in subsequent layers.
* Additional layers must be specified to fill the plot. 

### Geoms

Here are some commonly used layers for plotting in ggplot: 

* geom_bar
* geom_histogram
* geom_freqpoly
* geom_boxplot
* geom_violin
* geom_dotplot
* geom_point
* geom_errorbar
* geom_smooth
* geom_line
* geom_area
* geom_abline (plus geom_hline and geom_vline)
* geom_text

### Aesthetics
Here are some commonly used aesthetic types that can be manipulated in ggplot: 

* color
* fill
* shape
* size
* transparency

### Plotting continuous variables over time: Scatterplot
```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
ggplot(EPAair, aes(x = Date, y = Ozone)) + 
  geom_point()
# geom_point creates a scatter plot

O3plot <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone))
print(O3plot)
# print() allows us to see the plot

# Fix this code
O3plot2 <- ggplot(EPAair) +
# geom_point(aes(x = Date, y = Ozone, color = "blue"))
  geom_point(aes(x = Date, y = Ozone), color = "blue")
print(O3plot2)
# We can't change the color within our asthetics function, we need to put it outside (commented out section was the original code)

# Add additional variables
PMplot <- 
  ggplot(EPAair, aes(x = month, y = PM2.5, shape = as.factor(year), color = Site.Name)) +
  geom_point()
print(PMplot)
# Colors can be defined based on a variable from our dataframe (in this case Site.Name) 
# as.factor() changes the integer to a binary type factor
# B/c were coloring based on site name, it is placed inside the aes(), not outside like when we're defining specific colors
# If we want to make month more normal, we have to change the class (to what though?)

# Separate plot with facets
PMplot.faceted <-
  ggplot(EPAair, aes(x = month, y = PM2.5, shape = as.factor(year))) +
  geom_point() +
  facet_wrap(vars(Site.Name), nrow = 3)
print(PMplot.faceted)
# facet_wrap splits the graph based on Site.Name and arranges the graphs in 3 rows

PMplot.faceted2 <-
  ggplot(EPAair, aes(x = month, y = PM2.5)) +
  geom_point() +
  facet_grid(Site.Name ~ year)
print(PMplot.faceted2)
# ~ means by, so it sorts the graphs split by Site.Name by year
# Order does matter, if we did year by Site.Name, then the output would be flipped

# Filter dataset within plot building
O3plot.Blackstone <- 
  ggplot(subset(EPAair, Site.Name == "Blackstone"), aes(x = Date, y = Ozone)) + 
  geom_point() +
  geom_line()
print(O3plot.Blackstone)

# Exercise: build your own scatterplots of PeterPaul.chem.nutrients

# 1. 
# Plot surface temperatures by day of year. 
ggplot(PeterPaul.chem.nutrients, depth = 0.00, aes(x = daynum, y = temperature_C, color = year4)) +
  geom_point()

# Color your points by year, and facet by lake in two rows.
ggplot(PeterPaul.chem.nutrients, depth = 0.00, aes(x = daynum, y = temperature_C, color = year4)) +
  geom_point() +
  facet_wrap(vars(lakename), nrow = 2)

#2. 
# Plot temperature by date. Color your points by depth.
ggplot(PeterPaul.chem.nutrients, aes(x = sampledate, y = temperature_C, color = depth)) +
  geom_point()

# Change the size of your point to 0.5
ggplot(PeterPaul.chem.nutrients, aes(x = sampledate, y = temperature_C, color = depth)) +
  geom_point(size = 0.5)

```
### Plotting the relationship between two continuous variables: Scatterplot

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
lightvsDO <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point()
print(lightvsDO)
# We've got some crazy outliers, so we need to adjust our axes to zoom into the area of interest

# Adjust axes
lightvsDOfixed <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point() +
  xlim(0, 250) +
  ylim(0, 20)
print(lightvsDOfixed)
# x and y lim() zoomed to the clustered area. We see from this plot that DO is more variable with less light (deeper water). DO is closer to zero between 0-50 irradiance

# Depth in the fields of limnology and oceanography is on a reverse scale
tempvsdepth <- 
  #ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth)) +
  ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth, color = daynum)) +
  geom_point() +
  scale_y_reverse()
print(tempvsdepth)
# scale_y_reverse is super helpful here for making the graph easier to read (depth values reversed)

NvsP <-
  ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, color = depth)) +
  geom_point() +
  geom_smooth(method = lm) #+
  #geom_abline(aes(slope = 16, intercept = 0), lty = 2)
print(NvsP)
# Gray points in the output are measurements that don't have an associated depth value
# geom_smooth creates a line with the lease amount of error with the dataset (line of best fit), but to get a linear fit, we need to specify that the method is a linear model. It also includes a 95% confidence interval, saying that 95% of the points follow that line.
# geom_abline allows you to draw a linear line with any specifications we want. Here, we're telling it make a line with a slope of 16. If we expect the ratio of N to P to be 16 to 1, we can use this to confirm our assumptions. To make it a dotted line, include lty = 2 (lty means line type)

# Exercise: Plot relationships between air quality measurements

# 1. 
# Plot AQI values for ozone by PM2.5, colored by site. 
O3vsPM <-
  ggplot(EPAair, aes(x = PM2.5, y = Ozone, color = Site.Name)) +
  geom_point()
print(O3vsPM)

# Add a line of best fit for the linear regression of these variables.
O3vsPM <-
  ggplot(EPAair, aes(x = PM2.5, y = Ozone, color = Site.Name)) +
  geom_point() +
  geom_smooth(method = lm)
print(O3vsPM)
```

### Plotting continuous vs. categorical variables

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
 # Barplot + error bars
PeterPaul.nutrient.summaries <- PeterPaul.nutrients.gathered %>%
  group_by(lakename, nutrient) %>%
  summarise(sd = sd(concentration), 
            mean = mean(concentration))
# To group by multiple factors, they need to be in the same function. (Doing them in seperate functions will just overwrite the first one)

Nutrientplot <- 
  ggplot(PeterPaul.nutrients.gathered) +
  geom_bar(aes(x = lakename, y = concentration, fill = nutrient), # why did we use fill?
           position = "dodge", stat = "summary", fun.y = "mean")             # what's happening here?
print(Nutrientplot)
# We use fill because if we just use color, it will color the border of the bars and leave the inside gray.
# position e dodge seperates the columns vs. stacking the columns

Nutrientplot2 <- 
  ggplot(PeterPaul.nutrient.summaries, aes(x = lakename, y = mean, fill = nutrient)) + #
  geom_bar(stat = "identity", position = "dodge") + # what does the stat command do?
  geom_errorbar(aes(ymin = mean-sd, ymax = mean+sd), # how do we specify error bars?
                 position = "dodge")
print(Nutrientplot2)
# stat = identity tells it to just take the value in the cell for the y value (mean). We use this because we already summarized the data
# geom_errorbar- we need to have mean and sd columns in our dataset to do this calculation
# This is a bad way to represent our data because the error bars are huge, and they go below 0 (negative concentrations). Moreover, science is trending away from error bars

# Are there more effective ways to produce summary stats for categories?

# Box and whiskers plot
Nutrientplot3 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_boxplot(aes(color = nutrient)) # Why didn't we use "fill"?
print(Nutrientplot3)
# Didn't use fill so we can see the mean bar
# This does a better job of displaying the distribution of points while still communicating the means and such

# Dot plot
Nutrientplot4 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_dotplot(aes(color = nutrient), binaxis = "y", binwidth = 1, 
               stackdir = "center", position = "dodge") #
print(Nutrientplot4)
# Dot plot places measurements of the same value next to each other, so wide parts of the plot show value where most readings are

# Violin plot
Nutrientplot5 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_violin(aes(color = nutrient)) #
print(Nutrientplot5)
# Super similar to a dot plot, but it uses polygons instead of dots

# Frequency polygons
# Using a tidy dataset
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients) +
  geom_freqpoly(aes(x = tn_ug), color = "black") +
  geom_freqpoly(aes(x = tp_ug), color = "darkblue") +
  geom_freqpoly(aes(x = nh34), color = "darkgray") +
  geom_freqpoly(aes(x = no23), color = "gray") +
  geom_freqpoly(aes(x = po4), color = "blue") 
print(Nutrientplot6)

# Using a gathered dataset
Nutrientplot7 <-   
  ggplot(PeterPaul.nutrients.gathered) +
  geom_freqpoly(aes(x = concentration, color = nutrient))
print(Nutrientplot7)

# Exercise: Plot distributions of AQI values for EPAair

# 1. 
# Create a bar chart plus standard deviation error bars for PM2.5, divided by year. 
# Create separate bars for each site. 
EPAair.plot1 <-
  ggplot(EPAair) +
  geom_bar(aes(x = year, y = PM2.5, fill = Site.Name),
           position = "dodge", stat = "summary", fun.y = "mean") +
  geom_errorbar(aes(ymin = (mean(PM2.5) - sd(PM2.5))), ymax = (mean(PM2.5) + sd(PM2.5)))
print(EPAair.plot1)
# This was done correctly, however to include error bars, we need to summarize the data
  
# 2. 
# Create a new plot that better depicts the distribution of PM2.5 concentrations. 
# Divide your graph by year and site.
EPAair.plot2 <-
  ggplot(EPAair, aes(x = year, y = PM2.5)) +
  geom_boxplot(aes(color = Site.Name))
print(EPAair.plot2)
```
